# docs/api/api-docs.yaml
openapi: '3.0.0'

info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8900/
    description: Generated server url
paths:
  /api/offer/cmd/list:
    post:
      tags:
        - offer
      description: lists all travel offers
      operationId: listTravelOffers
      responses:
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: string
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDtoTravelOffersDetailsDto'
  /api/offer/cmd/add:
    post:
      tags:
        - offer
      description: adds a travel offer
      operationId: addTravelOffer
      responses:
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: string
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDtoTravelOffersDetailsDto'
  /api/offer/cmd/change:
    post:
      tags:
        - offer
      description: changes a travel offer
      operationId: changeTravelOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTravelOfferQueryDto"
      responses:
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: string
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDtoTravelOffersDetailsDto'
  /api/offer/cmd/search:
    post:
      tags:
        - offer
      description: searches travel offers
      operationId: searchTravelOffers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTravelOffersQueryDto"
      responses:
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: string
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDtoTravelOffersDetailsDto'
components:
  schemas:
    ChangeTravelOfferQueryDto:
      type: object
      properties:
        index:
          type: number
    SearchTravelOffersQueryDto:
      type: object
      properties:
        name:
          type: string
        nights:
          type: number
    PageDtoTravelOffersDetailsDto:
      type: object
      properties:
        elements:
          type: array
          items:
            $ref: '#/components/schemas/TravelOfferDetailsDto'
        totalCount:
          type: integer
      required:
        - elements
        - totalCount
    TravelOfferDetailsDto:
      type: object
      properties:
        id:
          type: number
        header:
          type: string
        content:
          type: array
          items:
            type: string
        from:
          type: string
          format: date-time
          description: Zeitpunkt der letzten Bearbeitung
        to:
          type: string
          format: date-time
          description: Zeitpunkt der letzten Bearbeitung
        price:
          type: number
      required:
        - header
        - content
        - from
        - to
        - price