# src/app/features/trip/api/model/changeTravelOfferQueryDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ChangeTravelOfferQueryDto {
    index?: number;
}