# src/app/features/trip/api/model/pageDtoTravelOffersDetailsDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {TravelOfferDetailsDto} from './travelOfferDetailsDto';

export interface PageDtoTravelOffersDetailsDto {
    elements: Array<TravelOfferDetailsDto>;
    totalCount: number;
}