# src/app/features/trip/api/model/travelOfferDetailsDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface TravelOfferDetailsDto {
    id?: number;
    header: string;
    content: Array<string>;
    /**
     * Zeitpunkt der letzten Bearbeitung
     */
    from: string;
    /**
     * Zeitpunkt der letzten Bearbeitung
     */
    to: string;
    price: number;
}